4/13

+++ new/database/migrations/2023_04_01_200647_create_be_histories_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('be_histories', function (Blueprint $table) {
             $table->id();
             $table->integer('userid');
-            $table->integer('big');
-            $table->integer('small');
+            $table->float('big');
+            $table->float('small');
             $table->integer('ticketno');
             $table->integer('total');
             $table->string('number');
